adding tests for scheduler
