--- services/director-v2/tests/integration/conftest.py
@@ -1,77 +1,0 @@
-# pylint:disable=no-value-for-parameter
-# pylint:disable=redefined-outer-name
-
-from random import randint
-from typing import Callable, Dict
-from uuid import uuid4
-
-import pytest
-import sqlalchemy as sa
-from models_library.projects import ProjectAtDB
-from pydantic.types import PositiveInt
-from simcore_postgres_database.models.projects import ProjectType, projects
-from simcore_postgres_database.models.users import UserRole, UserStatus, users
-from sqlalchemy import literal_column
-
-
-@pytest.fixture
-def user_id() -> PositiveInt:
-    return randint(0, 10000)
-
-
-@pytest.fixture
-def user_db(postgres_db: sa.engine.Engine, user_id: PositiveInt) -> Dict:
-    with postgres_db.connect() as con:
-        result = con.execute(
-            users.insert()
-            .values(
-                id=user_id,
-                name="test user",
-                email="test@user.com",
-                password_hash="testhash",
-                status=UserStatus.ACTIVE,
-                role=UserRole.USER,
-            )
-            .returning(literal_column("*"))
-        )
-
-        user = result.first()
-
-        yield dict(user)
-
-        con.execute(users.delete().where(users.c.id == user["id"]))
-
-
-@pytest.fixture
-def project(postgres_db: sa.engine.Engine, user_db: Dict) -> Callable:
-    created_project_ids = []
-
-    def creator(**overrides) -> ProjectAtDB:
-        project_config = {
-            "uuid": uuid4(),
-            "name": "my test project",
-            "type": ProjectType.STANDARD.name,
-            "description": "my test description",
-            "prj_owner": user_db["id"],
-            "access_rights": {"1": {"read": True, "write": True, "delete": True}},
-            "thumbnail": "",
-            "workbench": {},
-        }
-        project_config.update(**overrides)
-        with postgres_db.connect() as con:
-            result = con.execute(
-                projects.insert()
-                .values(**project_config)
-                .returning(literal_column("*"))
-            )
-
-            project = ProjectAtDB.parse_obj(result.first())
-            created_project_ids.append(project.uuid)
-            return project
-
-    yield creator
-
-    # cleanup
-    with postgres_db.connect() as con:
-        for pid in created_project_ids:
-            con.execute(projects.delete().where(projects.c.uuid == str(pid)))
